Rework github tag tool provider to use context
